+++ man-roxygen/transform-sd-args.R
@@ -1,4 +1,4 @@
 #' @param .how The function(s) or function call(s) that will perform the transformation. If many,
 #'   a list should be used, either with `list()` or `.()`. If the list is named, the names will be
-#'   used for the new columns' names.
+#'   used for the new columns' names. Lambdas specified as formulas are supported.
 #' @param ... Possibly more arguments for *all* functions/calls in `.how`.